refactor: username
