+++ src/app/[username]/components/Blogs.tsx
@@ -2,11 +2,16 @@
 
 import { BlogCard } from '@/components/blog/cards/BlogCard';
 import { BlogCardListSkeleton } from '@/components/skeletons/blogSkeleton';
-import useAuth from '@/hooks/auth/useAuth';
 import useGetPublishedBlogByUsername from '@/hooks/blog/useGetPublishedBlogByUsername';
+import { IUser } from '@/services/models/user';
 
-export const Blogs = ({ username }: { username: string }) => {
-  const { data: session, isSuccess } = useAuth();
+export const Blogs = ({
+  username,
+  user,
+}: {
+  username: string;
+  user?: IUser;
+}) => {
   const { blogs, isLoading, isError } = useGetPublishedBlogByUsername(username);
 
   if (isError)
@@ -34,8 +39,8 @@ export const Blogs = ({ username }: { username: string }) => {
               <BlogCard
                 key={blog?.blog_id}
                 blog={blog}
-                isAuthenticated={isSuccess}
-                modificationEnable={session?.username === username}
+                isAuthenticated={!!user}
+                modificationEnable={user?.username === username}
               />
             );
           })
