+++ src/app/[username]/components/NavigationTabs.tsx
@@ -1,11 +1,13 @@
-'use client';
-
 import { TabsList, TabsTrigger } from '@/components/ui/tabs';
-import useAuth from '@/hooks/auth/useAuth';
+import { IUser } from '@/services/models/user';
 
-export const NavigationTabs = ({ username }: { username: string }) => {
-  const { data } = useAuth();
-
+export const NavigationTabs = ({
+  username,
+  user,
+}: {
+  username: string;
+  user?: IUser;
+}) => {
   return (
     <TabsList className='flex justify-center gap-0'>
       <TabsTrigger value='blogs'>
@@ -16,7 +18,7 @@ export const NavigationTabs = ({ username }: { username: string }) => {
         <div className='mt-1 h-[1px] w-0 bg-brand-orange group-data-[state=active]:w-full transition-all' />
       </TabsTrigger>
 
-      {data?.username === username && (
+      {user?.username === username && (
         <TabsTrigger value='drafts'>
           <p className='px-3 font-dm_sans opacity-80 group-hover:opacity-100 group-data-[state=active]:opacity-100'>
             Drafts
