+++ src/app/[username]/page.tsx
@@ -1,36 +1,27 @@
-import type { Metadata } from 'next';
+'use client';
 
 import { Tabs, TabsContent } from '@/components/ui/tabs';
+import useAuth from '@/hooks/auth/useAuth';
 
 import { Blogs } from './components/Blogs';
 import { Drafts } from './components/Drafts';
 import { NavigationTabs } from './components/NavigationTabs';
 
-export async function generateMetadata({
-  params,
-}: {
-  params: { username: string };
-}): Promise<Metadata> {
-  const username = params.username;
-
-  return {
-    title: `@${username}`,
-  };
-}
-
 const ProfilePage = ({ params }: { params: { username: string } }) => {
+  const { data } = useAuth();
+
   return (
     <div>
       <Tabs defaultValue='blogs' className='space-y-6 md:space-y-8'>
-        <NavigationTabs username={params.username} />
+        <NavigationTabs username={params.username} user={data} />
 
         <div className='w-full'>
           <TabsContent className='w-full' value='blogs'>
-            <Blogs username={params.username} />
+            <Blogs username={params.username} user={data} />
           </TabsContent>
 
           <TabsContent className='w-full' value='drafts'>
-            <Drafts username={params.username} />
+            <Drafts username={params.username} user={data} />
           </TabsContent>
         </div>
       </Tabs>
